1. refactor to only extract fixture into parent class 2. remove apache-common Pair dependency
