+++ shardingsphere-proxy/shardingsphere-proxy-bootstrap/src/test/java/org/apache/shardingsphere/proxy/init/impl/AbstractBootstrapInitializerTest.java
@@ -20,17 +20,10 @@ package org.apache.shardingsphere.proxy.init.impl;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.SneakyThrows;
-import org.apache.commons.lang3.tuple.Pair;
-import org.apache.shardingsphere.infra.auth.Authentication;
-import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.infra.config.properties.ConfigurationProperties;
 import org.apache.shardingsphere.infra.context.SchemaContexts;
-import org.apache.shardingsphere.infra.context.schema.DataSourceParameter;
 import org.apache.shardingsphere.proxy.config.ProxyConfiguration;
 import org.apache.shardingsphere.proxy.config.YamlProxyConfiguration;
-import org.apache.shardingsphere.sharding.api.config.rule.ShardingTableRuleConfiguration;
-import org.apache.shardingsphere.sharding.api.config.strategy.sharding.ShardingStrategyConfiguration;
-import org.apache.shardingsphere.sharding.api.config.strategy.sharding.StandardShardingStrategyConfiguration;
 import org.apache.shardingsphere.transaction.context.TransactionContexts;
 import org.junit.Before;
 import org.junit.Ignore;
@@ -40,15 +33,10 @@ import org.mockito.Mockito;
 import java.io.IOException;
 import java.net.Socket;
 import java.net.UnknownHostException;
-import java.util.Collection;
-import java.util.Map;
 import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
-import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
@@ -73,41 +61,11 @@ public abstract class AbstractBootstrapInitializerTest {
         prepareSpecifiedInitializer();
     }
     
-    protected void doAssertGetProxyConfiguration() {
-        beforeAssertGetProxyConfiguration();
-        YamlProxyConfiguration yamlConfig = makeProxyConfiguration();
-        ProxyConfiguration actual = initializer.getProxyConfiguration(yamlConfig);
-        assertNotNull(actual);
-        assertProxyConfiguration(actual);
-        afterAssertGetProxyConfiguration();
-    }
-    
-    protected void beforeAssertGetProxyConfiguration() {
-    
-    }
-    
-    protected void afterAssertGetProxyConfiguration() {
-    
-    }
-    
-    protected YamlProxyConfiguration makeProxyConfiguration() {
-        return null;
-    }
-    
-    protected void assertProxyConfiguration(final ProxyConfiguration actual) {
-        assertSchemaDataSources(actual.getSchemaDataSources());
-        assertSchemaRules(actual.getSchemaRules());
-        assertAuthentication(actual.getAuthentication());
-        assertProps(actual.getProps());
-    }
-    
     protected void doEnvironmentPrepare() {
     
     }
     
-    protected void prepareSpecifiedInitializer() {
-    
-    }
+    protected abstract void prepareSpecifiedInitializer();
     
     @SneakyThrows
     @Test
@@ -115,16 +73,6 @@ public abstract class AbstractBootstrapInitializerTest {
         new Thread(this::triggerAbstractBootstrapInitializerInit).start();
         TimeUnit.SECONDS.sleep(5);
         assertTrue(isAvailable());
-    }
-    
-    protected Pair<SchemaContexts, SchemaContexts> makeDecoratedSchemaContexts() {
-        SchemaContexts original = mock(SchemaContexts.class);
-        return Pair.of(original, initializer.decorateSchemaContexts(original));
-    }
-    
-    protected Pair<TransactionContexts, TransactionContexts> makeDecoratedTransactionContexts() {
-        TransactionContexts original = mock(TransactionContexts.class);
-        return Pair.of(original, initializer.decorateTransactionContexts(original));
     }
     
     @SneakyThrows
@@ -151,63 +99,6 @@ public abstract class AbstractBootstrapInitializerTest {
             portFree = false;
         }
         return portFree;
-    }
-
-    protected void assertSchemaDataSources(final Map<String, Map<String, DataSourceParameter>> actual) {
-    
-    }
-    
-    protected void assertDataSourceParameter(final DataSourceParameter actual) {
-        assertThat(actual.getUrl(), is("jdbc:fixturedb:xxx"));
-        assertThat(actual.getUsername(), is("root"));
-        assertThat(actual.getPassword(), is("pwd"));
-        assertTrue(actual.isReadOnly());
-        assertThat(actual.getConnectionTimeoutMilliseconds(), is(1000L));
-        assertThat(actual.getIdleTimeoutMilliseconds(), is(2000L));
-        assertThat(actual.getMaxLifetimeMilliseconds(), is(4000L));
-        assertThat(actual.getMaxPoolSize(), is(20));
-        assertThat(actual.getMinPoolSize(), is(10));
-    }
-    
-    protected void assertSchemaRules(final Map<String, Collection<RuleConfiguration>> actual) {
-        assertThat(actual.size(), is(1));
-        String key = fetchKey();
-        assertTrue(actual.containsKey(key));
-        Collection<RuleConfiguration> ruleConfigurations = actual.get(key);
-        assertThat(ruleConfigurations.size(), is(1));
-        assertRuleConfiguration(ruleConfigurations.iterator().next());
-    }
-    
-    protected String fetchKey() {
-        return null;
-    }
-    
-    protected void assertRuleConfiguration(final RuleConfiguration actual) {
-    
-    }
-    
-    protected void assertShardingTableRules(final Collection<ShardingTableRuleConfiguration> shardingTableRules) {
-        assertNotNull(shardingTableRules);
-        assertThat(shardingTableRules.size(), is(1));
-        ShardingTableRuleConfiguration shardingTableRule = shardingTableRules.iterator().next();
-        assertNotNull(shardingTableRule);
-        assertThat(shardingTableRule.getLogicTable(), is("t_order"));
-        assertThat(shardingTableRule.getActualDataNodes(), is("ds_${0..1}.t_order_${0..1}"));
-        assertShardingStrategy(shardingTableRule.getTableShardingStrategy(), "order_id", "t_order_inline");
-        assertShardingStrategy(shardingTableRule.getDatabaseShardingStrategy(), "user_id", "database_inline");
-        assertNull(shardingTableRule.getKeyGenerateStrategy());
-    }
-    
-    private void assertShardingStrategy(final ShardingStrategyConfiguration shardingStrategy, final String shardingColumn, final String shardingAlgorithmName) {
-        assertNotNull(shardingStrategy);
-        assertThat(shardingStrategy, instanceOf(StandardShardingStrategyConfiguration.class));
-        StandardShardingStrategyConfiguration standardShardingStrategy = (StandardShardingStrategyConfiguration) shardingStrategy;
-        assertThat(standardShardingStrategy.getShardingColumn(), is(shardingColumn));
-        assertThat(standardShardingStrategy.getShardingAlgorithmName(), is(shardingAlgorithmName));
-    }
-    
-    protected void assertAuthentication(final Authentication actual) {
-    
     }
     
     protected void assertProps(final Properties actual) {
